add shared library / dll support to the build
- for Linux, create a shared library
- for the Windows MinGW build, create a DLL

Note: the thirdparty static libraries that are compiled by the build are
folded into the resulting MuPDF library so it can be installed without
conflicts if the system version of those libraries differs. The symbols
defined by those thirdparty static libraries are however not exported by
the resulting MuPDF library (again to avoid conflicts when building an
application that needs to use a different versions of the same
thirdparty library).